drm/i915: Introduce intel_schedule_vblank_job()
Various places in the driver need the ability to schedule actions to run
on a future vblank (updating watermarks, unpinning buffers, etc.).  The
long term goal is to add such a mechanism in the DRM core that can be
shared by all drivers.  Paulo Zanoni has already written some
preliminary patches to support this, but his work will probably take
some time to polish and get merged since it needs to untangle the DRM
core's vblank infrastructure.  This patch takes a simpler approach and
just adds an i915-specific mechanism that can be used in the interim
since we have an immediate need for watermark updates.

Cc: Paulo Zanoni <przanoni@gmail.com>
Cc: Ville Syrjälä <ville.syrjala@linux.intel.com>
Signed-off-by: Matt Roper <matthew.d.roper@intel.com>