drm/i915: Clear crtc atomic flags at beginning of transaction
Once we have full atomic modeset, these kind of flags should be in a
real intel_crtc_state that's tracked properly.  In the meantime, make
sure we clear out any old flags at the beginning of a transaction so
that we don't wind up seeing leftover flags from old transactions that
were checked, but never went to the commit step.

A simple memset would have done here, but I expect there to be a few
more things that we *don't* want to clear that get added into this
structure before we're ready to kill off and roll everything into the
CRTC state.

Signed-off-by: Matt Roper <matthew.d.roper@intel.com>