drm/atomic/helper: Allow duplication of in-flight states
Drivers may wish to duplicate in-flight states internally, but the
__drm_atomic_helper_{crtc,plane}_duplicate_state helper functions can
only be used to make a copy of state that has already been committed to
obj->state.  Let's update the helpers to take any existing state object
as their first parameter rather than the DRM object pointer.  We can
update handful of existing callsites to just pass obj->state for it.

Cc: dri-devel@lists.freedesktop.org
Signed-off-by: Matt Roper <matthew.d.roper@intel.com>