drm/i915: Extract plane dumping from intel_dump_pipe_config()
The plane information isn't important in most of the places where we
dump the pipe config; the one place where it would be valuable is during
initial HW state readout, but currently we dump the pipe config before
FB's are reconstructed, so the data dumped is misleading anyway.  Even
worse, in places where we're dumping an in-flight pipe config, we're
still dumping the already-committed plane state, which is
confusing/misleading.

Extract the plane dumping to its own function that can be called in the
specific place(s) where we actually want to see this information.

A few other changes are made while extracting this dumping logic:
 * show planes that have an FB, but are invisible (e.g., because
   they're offscreen or hidden by other planes) as "not visible" rather
   than "enabled."
 * display the FB format as a human-readable string rather than a hex
   format code
 * List the actual plane type ("Primary," "Overlay," or "Cursor") rather
   than just "STANDARD" or "CURSOR."

Suggested-by: Maarten Lankhorst <maarten.lankhorst@linux.intel.com>
Cc: Maarten Lankhorst <maarten.lankhorst@linux.intel.com>
Signed-off-by: Matt Roper <matthew.d.roper@intel.com>